requirements.txt
python-telegram-bot==21.6
gspread==6.1.4
google-auth==2.33.0
google-api-python-client==2.143.0
python-dateutil==2.9.0.post0